Created a YOLOv5 data fetching/training pipeline.
